requirements.txt
Django==4.2
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1